This is the rough idea, the information on slides 1- 20 will stay relatively the same, possibly just reworded
This is what I've got so far guys. I just want this to look the best possible and I got screwed by work, so I am going to continue working on it tonight and tomorrow for class. However this should be most of the information we've discussed, with a few changes here and there. Tell me what you think and what other stuff I can keep adding. I plan to add a revised statement of work with clearly defined goals and basic business rules